providers: move `get_provider()` to providers.py
