+++ tests/unit/parts/test_lifecycle.py
@@ -67,7 +67,7 @@ def project_vars(mocker):
 @pytest.fixture()
 def mock_provider(mocker, mock_instance, fake_provider):
     _mock_provider = Mock(wraps=fake_provider)
-    mocker.patch("snapcraft.providers.get_provider", return_value=_mock_provider)
+    mocker.patch("snapcraft.parts.lifecycle.get_provider", return_value=_mock_provider)
     yield _mock_provider
